directory: split files and fix comments
